instructions
This is a distributed bellman ford algorithm, which can be run on 4 terminals or 4 different computers. This algorithm dynamically changes the routing protocol if there're any changes in any part of the network. And finally any node in this network will get a routing table with the shortest path to any other node in the network and the next node to take.
How to Run
1. make
2. each terminal type in a line
java BellmanFordNode 127.0.0.1 1111 3000 127.0.0.1 2222 5 127.0.0.1 3333 2
java BellmanFordNode 127.0.0.1 2222 3000 127.0.0.1 1111 5 127.0.0.1 4444 8
java BellmanFordNode 127.0.0.1 3333 3000 127.0.0.1 1111 2 127.0.0.1 4444 20
java BellmanFordNode 127.0.0.1 4444 3000 127.0.0.1 3333 20 127.0.0.1 2222 8
3.Other commands work as programming requirement

Extra Functionalities
1.Poisoned reverse from BellmanFordNode.java line 247
2.Graphical user interface to show routing table.